Restructuring, using useSetAtom, and renaming queryKey
