+++ frontend/src/modules/Pvt/settings/components/DependentVariableSelector/index.ts
